services/users/requirements.txt
Flask==0.12.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.1
Flask-Testing==0.6.2
gunicorn==19.7.1
coverage==4.4.1
flake8===3.4.1
flask-cors==3.0.3
flask-migrate==2.1.1
